refactor(retrieval): move the remaining tools into tools/
grounded-retrieval relocates from packages/tools — it is a retrieval
policy, not a vertical capability, and holding it in tools forced a
cross-brick import for every consumer. A re-export keeps the old path
alive for brand-voice/company-context/claim-evidence/persona and the
pipelines' architecture rule ("no feature calls the RAG port directly").

The cost-aware RLM search service moves out of documentQ&A wholesale:
its only app dependencies turned out to be compositions of llm-package
primitives (createEmbeddingModel + resolveEmbeddingIndex) and the store
client, so it lands in tools/rlm-search with zero injected seams. The
Q&A services barrel and AIQueryRLM route re-point to the package.

Every tool folder now carries the convention: README, index.ts surface,
implementation beside it.
