+++ packages/retrieval/src/tools/rlm-search/index.ts
@@ -1,0 +1,9 @@
+export {
+    performRLMSearch,
+    getDocumentOverviewForPlanning,
+    getDocumentOverviewsBatch,
+    getDocumentStructureTree,
+    getSectionsByPath,
+    type RLMSearchOptions,
+    type RLMSearchResult,
+} from "./rlm-search";
